requirements.txt
Pillow==8.4.0
matplotlib==3.4.3
numpy==1.21.2
tensorflow==2.6.0
torch==1.9.1
torchsummary==1.5.1
torchvision==0.10.1
tensorflow-datasets==4.5.0